debug

+++ replyMain.py
@@ -25,7 +25,7 @@ group.add_argument("-cuda", type=int, default=-1, help="use GPU")
 group.add_argument("-double", action='store_true', help="use float64")
 group.add_argument("-lr", type=float, default=0.001, help="learning rate")
 group.add_argument("-savePeriod", type=int, default=10, help="")
-group.add_argument("-alpha", type=int, default=1, help="")
+group.add_argument("-alpha", type=float, default=1, help="")
 
 group = parser.add_argument_group('network parameters')
 group.add_argument("-load", action='store_true', help="if load from folder")
